requirements.txt
asgiref==3.11.0
Django==5.2.8
psycopg2-binary==2.9.11
sqlparse==0.5.4
psycopg2-binary==2.9.11
djangorestframework==3.16.1
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
pytesseract==0.3.13
Pillow==12.0.0
python-decouple==3.8
pytesseract==0.3.13